+++ ChangeLog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [Unreleased]
+### Fixed
+
+- `init` command generating a project skeleton as expected by `render`.
 
 [0.12.0] - 2025-03-01
 ### Added